Fixes filtering of duplicates in the reducer
