+++ lib/tags/index.js
@@ -11,9 +11,20 @@ export default class TagContainer extends Component {
     }
   }
   render () {
-    const { tags, removeTag, onSearchButtonClick, setSearchString } = this.props;
+    const {
+      tags,
+      removeTag,
+      onSearchButtonClick,
+      setSearchString
+    } = this.props;
+    console.log('remove tag', removeTag);
     var tagComponents = tags.map(function (tag) {
-      return <Tag key={tag.displayName} displayName={tag.displayName} colour={tag.colour} removeTag={removeTag}/>;
+      return <Tag
+        key={tag.displayName}
+        displayName={tag.displayName}
+        colour={tag.colour}
+        removeTag={removeTag}
+      />;
     });
     return (
       <div className='tagWrapper'>
